+++ src/test/java/com/github/vy/fibertest/Util.java
@@ -14,13 +14,13 @@ class Util {
     static void testRingBenchmark(
             final int workerCount,
             final int ringSize,
-            final int[] sequences) {
+            final Integer[] sequences) {
         final int offset = workerCount - ringSize % workerCount;
         for (int i = 0; i < workerCount; i++)
             try {
                 assertEquals(
                         "sequence returned by Worker#" + i,
-                        -((offset + i) % workerCount), sequences[i]);
+                        -((offset + i) % workerCount), sequences[i].intValue());
             } catch (AssertionError ae) {
                 log.trace("sequences[] = {}", Arrays.toString(sequences));
                 throw ae;
